requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
attrs==23.2.0
bidict==0.22.1
click==8.1.7
colorama==0.4.6
exceptiongroup==1.2.0
fastapi==0.109.2
frozenlist==1.4.1
h11==0.14.0
httptools==0.6.1
idna==3.6
Jinja2==3.1.3
MarkupSafe==2.1.5
multidict==6.0.5
pydantic==2.6.1
pydantic_core==2.16.2
python-dotenv==1.0.1
python-engineio==4.9.0
python-socketio==5.11.1
PyYAML==6.0.1
simple-websocket==1.0.0
sniffio==1.3.0
starlette==0.36.3
typing_extensions==4.9.0
uvicorn==0.27.1
watchfiles==0.21.0
websockets==12.0
wsproto==1.2.0
yarl==1.9.4
